# common-platform-customer-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-customer-mgmt
  description: Customer Management Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/customers
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/parties/{partyId}/relationships/{relationshipId}:
    get:
      tags:
        - Party Relationships
      summary: Get party relationship by ID
      description: Retrieve a specific relationship associated with a party by its unique identifier
      operationId: getPartyRelationshipById
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: relationshipId
          in: path
          description: Unique identifier of the party relationship
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Party relationship found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartyRelationshipDTO'
        '404':
          description: Party relationship or party not found
    put:
      tags:
        - Party Relationships
      summary: Update party relationship
      description: Update an existing relationship associated with a party
      operationId: updatePartyRelationship
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: relationshipId
          in: path
          description: Unique identifier of the party relationship
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyRelationshipDTO'
        required: true
      responses:
        '200':
          description: Party relationship successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartyRelationshipDTO'
        '400':
          description: Invalid party relationship data
        '404':
          description: Party relationship or party not found
    delete:
      tags:
        - Party Relationships
      summary: Delete party relationship
      description: Delete a relationship associated with a party from the system
      operationId: deletePartyRelationship
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: relationshipId
          in: path
          description: Unique identifier of the party relationship
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Party relationship successfully deleted
        '404':
          description: Party relationship or party not found
...
  /api/v1/parties/{partyId}/relationships:
    post:
      tags:
        - Party Relationships
      summary: Create party relationship
      description: Create a new relationship between the specified party and another party
      operationId: createPartyRelationship
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyRelationshipDTO'
        required: true
      responses:
        '201':
          description: Party relationship successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartyRelationshipDTO'
        '400':
          description: Invalid party relationship data
        '404':
          description: Party not found
  /api/v1/parties/{partyId}/relationships/filter:
    post:
      tags:
        - Party Relationships
      summary: Filter party relationships
      description: Retrieve a paginated list of relationships associated with a specific party based on filtering criteria
      operationId: filterPartyRelationships
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestPartyRelationshipDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered party relationships
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
        '404':
          description: Party not found
...
components:
  schemas:
...
    PartyRelationshipDTO:
      required:
        - fromPartyId
        - relationshipTypeId
        - toPartyId
      type: object
      properties:
        partyRelationshipId:
          type: string
          format: uuid
          readOnly: true
          example: "123e4567-e89b-12d3-a456-426614174001"
        fromPartyId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        toPartyId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174002"
        relationshipTypeId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174003"
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        active:
          type: boolean
        notes:
          maxLength: 500
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
    FilterRequestPartyRelationshipDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/PartyRelationshipDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for party relationships
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...